修改外來鍵資料型態、建立students和teachers跟users的關聯

+++ database/migrations/2022_01_02_093818_create_electives_table.php
@@ -15,8 +15,8 @@ class CreateElectivesTable extends Migration
     {
         Schema::create('electives', function (Blueprint $table) {
             $table->id();
-            $table->integer('student_id');
-            $table->integer('course_id');
+            $table->unsignedBigInteger('student_id');
+            $table->unsignedBigInteger('course_id');
             $table->integer('score');
             $table->timestamps();
         });
