+++ database/migrations/2022_01_02_094159_create_homework_table.php
@@ -15,7 +15,7 @@ class CreateHomeworkTable extends Migration
     {
         Schema::create('homework', function (Blueprint $table) {
             $table->id();
-            $table->integer('course_id');
+            $table->unsignedBigInteger('course_id');
             $table->timestamps();
         });
     }
